remove stragglers
