+++ apps/studio/src/common/interfaces/IConnection.ts
@@ -1,6 +1,4 @@
-import { AzureAuthOptions, LibSQLOptions, RedshiftOptions } from "../appdb/models/saved_connection"
-import { BigQueryOptions } from "../appdb/models/saved_connection"
-import { CassandraOptions } from "../appdb/models/saved_connection"
+import { AzureAuthOptions, BigQueryOptions, CassandraOptions, LibSQLOptions, RedshiftOptions } from "@/lib/db/types"
 import { Transport } from "../transport"
 
 const ConnectionTypes = ['sqlite', 'sqlserver', 'redshift', 'cockroachdb', 'mysql', 'postgresql', 'mariadb', 'cassandra', 'oracle', 'bigquery', 'firebird', 'tidb', 'libsql'] as const
