+++ apps/studio/src/lib/db/authentication/azure.ts
@@ -4,26 +4,10 @@ import { wait } from '@shared/lib/wait';
 import rawLog from 'electron-log';
 import { TokenCache } from '@/common/appdb/models/token_cache';
 import globals from '@/common/globals';
+import { AzureAuthType } from '../types';
 
 const log = rawLog.scope('auth/azure');
 
-export enum AzureAuthType {
-  Default, // This actually may not work at all, might need to just give up on it
-  Password,
-  AccessToken,
-  MSIVM,
-  ServicePrincipalSecret
-}
-
-// supported auth types that actually work :roll_eyes: default i'm looking at you
-export const AzureAuthTypes = [
-  // Can't have 2FA, kinda redundant now
-  // { name: 'Password', value: AzureAuthType.Password },
-  { name: 'Azure AD SSO', value: AzureAuthType.AccessToken },
-  // This may be reactivated when we move to client server architecture
-  // { name: 'MSI VM', value: AzureAuthType.MSIVM },
-  { name: 'Azure Service Principal Secret', value: AzureAuthType.ServicePrincipalSecret }
-];
 
 type CloudTokenResponse = {
   cloud_token: CloudToken,
