+++ apps/studio/src/plugins/SettingsPlugin.ts
@@ -1,11 +1,11 @@
-import { UserSetting } from "@/common/appdb/models/user_setting"
+import Vue from 'vue';
 
 
 
 export const SettingsPlugin = {
 
   async get(key: string, defaultValue?: any) {
-    const result = await UserSetting.findOne({key})
+    const result = await Vue.prototype.$util.send('appdb/setting/get', { key });
     if (result) {
       return result.value
     }
@@ -13,21 +13,8 @@ export const SettingsPlugin = {
   },
 
   async set(key: string, value: string) {
-    const existing = await UserSetting.findOne({key})
-    if (existing) {
-      existing.userValue = value
-      await existing.save()
-    } else {
-      const nu = new UserSetting()
-      nu.key = key
-      nu.defaultValue = ''
-      nu.valueType = 0
-      nu.userValue = value
-      await nu.save()
-    }
+    await Vue.prototype.$util.send('appdb/setting/set', { key, value });
   }
-
-
 }
